кароче чето делал не помню че

+++ module/MxmApi/src/V1/Rest/File/FileResource.php
@@ -16,6 +16,8 @@ use ZF\MvcAuth\Identity\AuthenticatedIdentity;
 use MxmRbac\Service\AuthorizationService;
 use MxmUser\Mapper\MapperInterface;
 use MxmUser\Exception\RecordNotFoundUserException;
+use Zend\Log\Logger;
+use Zend\Stdlib\ErrorHandler;
 
 class FileResource extends AbstractResourceListener
 {
@@ -49,13 +51,19 @@ class FileResource extends AbstractResourceListener
      */
     protected $userMapper;
 
+    /**
+     * @var Zend\Log\Logger
+     */
+    protected $logger;
+
     public function __construct(
         TableGateway $tableGateway,
         \DateTimeInterface $datetime,
         Config $config,
         Response $response,
         AuthorizationService $authorizationService,
-        MapperInterface $mapper
+        MapperInterface $mapper,
+        Logger $logger
     ){
         $this->tableGateway = $tableGateway;
         $this->datetime = $datetime;
@@ -63,6 +71,7 @@ class FileResource extends AbstractResourceListener
         $this->response = $response;
         $this->authorizationService = $authorizationService;
         $this->userMapper = $mapper;
+        $this->logger = $logger;
     }
     /**
      * Create a resource
@@ -83,11 +92,19 @@ class FileResource extends AbstractResourceListener
         if ($identity instanceof AuthenticatedIdentity) {
             $authenticatedIdentity = $identity->getAuthenticationIdentity();
             if (! $authenticatedIdentity) {
-                unlink($file['tmp_name']);
+                ErrorHandler::start();
+                $test = unlink($file['tmp_name']);
+                $error = ErrorHandler::stop();
+                if (! $test) {
+                    $this->logger->err('FileResource. Unauthenticated attempt to create file. Cannot remove tmp file ' . $file['tmp_name'] . '. ' . $error);
+                } else {
+                    $this->logger->warn('FileResource. Unauthenticated attempt to create file. Tmp file was removed: ' . $file['tmp_name']);
+                }
 
-                return new ApiProblem(401, 'Unauthorized');
+                return new ApiProblem(401, 'Unauthenticated');
             }
         } else {
+
             return new ApiProblem(401, 'Can not get identity');
         }
 
@@ -99,7 +116,14 @@ class FileResource extends AbstractResourceListener
 
         $this->authorizationService->setIdentity($currentUser);
         if (! $this->authorizationService->isGranted('create.file.rest')) {
-            unlink($file['tmp_name']);
+            ErrorHandler::start();
+            $test = unlink($file['tmp_name']);
+            $error = ErrorHandler::stop();
+            if (! $test) {
+                $this->logger->err('FileResource. Unauthorized attempt to create file. Cannot remove tmp file ' . $file['tmp_name'] . '. ' . $error);
+            } else {
+                $this->logger->warn('FileResource. Unauthorized attempt to create file. Remove tmp file ' . $file['tmp_name']);
+            }
 
             return new ApiProblem(403, 'Forbidden. Permission create.file.rest is required.');
         }
@@ -118,7 +142,14 @@ class FileResource extends AbstractResourceListener
 
         $resultSet = $this->tableGateway->select(['id' => $id]);
         if (0 === count($resultSet)) {
-            unlink($file['tmp_name']);
+            ErrorHandler::start();
+            $test = unlink($file['tmp_name']);
+            $error = ErrorHandler::stop();
+            if (! $test) {
+                $this->logger->err('FileResource. Insert operation failed or did not result in new row. Cannot remove tmp file ' . $file['tmp_name'] . '. ' . $error);
+            } else {
+                $this->logger->warn('FileResource. Insert operation failed or did not result in new row. Remove tmp file ' . $file['tmp_name']);
+            }
 
             return new ApiProblem(500, 'Insert operation failed or did not result in new row');
         }
@@ -171,11 +202,20 @@ class FileResource extends AbstractResourceListener
             return new ApiProblem(404, 'File not found.');
         }
 
-        unlink($path);
+        ErrorHandler::start();
+        $test = unlink($path);
+        $error = ErrorHandler::stop();
+        if (! $test) {
+            $this->logger->err('FileResource. Cannot remove file ' . $path . '. ' . $error);
+
+            return new ApiProblem(401, 'Cannot remove file.');
+        }
 
         $result = $this->tableGateway->delete(['id' => $id]);
 
         if (!$result) {
+            $this->logger->err('FileResource. Cannot remove file record from db ' . $path);
+
             return false;
         }
